Avoid writing same filenames in different directories
Right now, when we have a new collection minor version, a new pdf is
created with the same name in the first "storage" in the rhaptos print
tool.  We sometimes end up having multiple files with the same name
because legacy doesn't use collection minor versions.

The way to fix this is by looking for the existing file and overwriting
the original file.

Change RhaptosPrintTool._getContainer to look in all containers for the
filename and return the container if found, otherwise return the first
container.